requirements.txt
openai>=1.50.0
anthropic>=0.40.0
google-genai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0

